pyproject.toml
[project]
name = "pbchs-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.0",
    "fpdf2>=2.8.5",
    "groq>=1.0.0",
    "langgraph>=1.0.5",
    "langsmith>=0.6.0",
    "psycopg2-binary>=2.9.11",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.14.3",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "sqlglot>=28.5.0",
]
